cloud_functions/schema_refresh/requirements.txt
functions-framework==3.*
google-cloud-bigquery>=3.11.0
langchain-google-vertexai>=2.0.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain>=0.3.0
pypdf>=3.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
